+++ src/features/distance/providers/DistanceProvider.tsx
@@ -2,16 +2,36 @@
 
 import { useDistance } from '@/features/distance/hooks/useDistance';
 import { createContextWithUseHook } from '@/util/createContextWithUseHook';
+import { getThresholdAdjusted } from '../util/getThresholdAdjusted';
+import { useGeolocationContext } from '@/features/geolocation/providers/GeolocationProvider';
+import { useEventContext } from '@/features/events/providers/EventProvider';
 
 const [DistanceContext, useDistanceContext] = createContextWithUseHook<{
   distance: number;
+  joinThreshold: number;
+  leaveThreshold: number;
   distancePending: boolean;
 }>('useDistanceContext can only be called within the context of a DistanceContext!');
 
 export function DistanceProvider({ children }: React.PropsWithChildren) {
+  const { position } = useGeolocationContext();
   const { distance, isPending } = useDistance();
+  const { event } = useEventContext();
+  const joinThreshold = getThresholdAdjusted(
+    5,
+    position?.coords.accuracy || 0,
+    event.position_accuracy
+  );
+
+  const leaveThreshold = getThresholdAdjusted(
+    15,
+    position?.coords.accuracy,
+    event.position_accuracy
+  );
+
   return (
-    <DistanceContext.Provider value={{ distance, distancePending: isPending }}>
+    <DistanceContext.Provider
+      value={{ distance, joinThreshold, leaveThreshold, distancePending: isPending }}>
       {children}
     </DistanceContext.Provider>
   );