+++ src/hooks/useTimeout.ts
@@ -3,13 +3,20 @@ import { useCallback, useEffect, useRef } from 'react';
 type TTimeoutContainer = {
   timeout: NodeJS.Timeout;
   start: number;
+  /**Optional function fired if the timeout is removed; either explicitly, or during unmount. */
+  cleanup?: () => void;
 };
 
 export function useTimeout() {
   const timeoutMap = useRef(new Map<string, TTimeoutContainer>()).current;
 
   /**Adds a new timeout for the given key. Will clear and replace the old if one exists.*/
-  const addTimeout = (key: string, cb: () => void | Promise<void>, t: number) => {
+  const addTimeout = (
+    key: string,
+    cb: () => void | Promise<void>,
+    t: number,
+    cleanup?: () => void
+  ) => {
     const ot = timeoutMap.get(key);
     if (ot) {
       clearTimeout(ot.timeout);
@@ -18,6 +25,7 @@ export function useTimeout() {
     timeoutMap.set(key, {
       timeout: nt,
       start: Date.now(),
+      cleanup,
     });
   };
 
@@ -25,6 +33,10 @@ export function useTimeout() {
     const t = timeoutMap.get(key);
     if (t) {
       clearTimeout(t.timeout);
+
+      if (t.cleanup) {
+        t.cleanup();
+      }
       timeoutMap.delete(key);
     }
   };
@@ -37,9 +49,8 @@ export function useTimeout() {
 
   useEffect(() => {
     return () => {
-      for (const [key, val] of timeoutMap) {
-        clearTimeout(val.timeout);
-        timeoutMap.delete(key);
+      for (const key of timeoutMap.keys()) {
+        removeTimeout(key);
       }
     };
   }, []);